backend-storarge/connected/requirements.txt
Django==1.4.2
psycopg2==2.4.5
dj-database-url==0.2.0
django-tastypie==0.9.11

